rename: type 폴더명 enums로 변경
