+++ src/main/java/com/dain_review/domain/campaign/model/entity/enums/Platform.java
@@ -1,4 +1,4 @@
-package com.dain_review.domain.campaign.model.type;
+package com.dain_review.domain.campaign.model.entity.enums;
 
 
 import lombok.AllArgsConstructor;
