+++ src/main/java/com/dain_review/domain/user/model/entity/enums/Gender.java
@@ -1,0 +1,6 @@
+package com.dain_review.domain.user.model.entity.enums;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
