+++ src/main/java/com/dain_review/domain/user/model/entity/enums/Role.java
@@ -1,4 +1,4 @@
-package com.dain_review.domain.user.model.type;
+package com.dain_review.domain.user.model.entity.enums;
 
 public enum Role {
     ROLE_INFLUENCER,
