+++ src/main/java/com/dain_review/domain/user/model/entity/enums/SnsType.java
@@ -1,4 +1,4 @@
-package com.dain_review.domain.user.model.type;
+package com.dain_review.domain.user.model.entity.enums;
 
 
 import lombok.AllArgsConstructor;
