--- src/main/java/com/dain_review/domain/user/model/type/Gender.java
@@ -1,6 +1,0 @@
-package com.dain_review.domain.user.model.type;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
